Removed mapfile cleanup workaround, implemented solution
We need to remove the mapping file for the session when it is closing
and the close script is called via pam_script, but not remove it when
the close is called from ending an sudo session, etc.

Check which processes are in the current session, and if any of them
are not one of sshd, sudo, login, su, or telnetd, then don't do the
cleanup (normally it will be a shell).

Cleanup won't happen if the user leaves jobs running when they logout
(e.g., via setsid), but that's fairly benign.  Even with very long
system uptimes, these are small files, and will not be a significant
issue.   At some future point, we might add garbage collection for
any session files found in the dbdir.