Do not use mapuser functionality with useradd,userdel,usermod
Ticket: CM-17450
Reviewed By: olson
Testing Done: ran programs with change

The useradd family will not work correctly with the mapuser/mapuid
functionality, and useradd provides no method to force creating
a user that already exists.

So check which program invoked us, using __progname (getprogname() could
also be used for non-glibc use), and return NOTFOUND immediately in that
case.

This is a major hack, but it's simple, and avoids a significant issue.

Unfortunately, the RADIUS protocol gives us no way to determine that
an account name is valid without also authenticating, and libnss plugins
do not have the ability to authenticate.