cleanup group mapname in nsswitch on remove; add priv user to sudo group
Needed to add sed code to remove mapname from nsswitch.conf group
search line, similar to passwd line.  Somehow forgot that when I
added the code to add it in postinst

I also somehow forgot to checkin the adduser line to add the
radius_priv_user to the sudo group, so fixed that too.