Changes and cleanups around the privileged user.
Ticket:  CM-19457
Reviewed By: nobody
Testing Done:  lots of logins, and login combinations

Validate that the mapuser database files are valid by using kill 0 on
the pids.  If not valid, try to unlink, and if we do, report it as a
DEBUG.  If we can't unlink (not root) report that we are skipping at INFO.

As part of that, don't count valid fields and read entire file.

Document in man page and config file that the mapped_priv_user account
is known to libpam-radius-auth as well, and must be updated in both
places if it changes.

Updated the public symbols (the getgr additions) from previous commit

Fixed some white space and line length issues.